move login functionality to login.js
